requirements-dev.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0
pytest-mock>=3.10.0
requests-mock>=1.11.0
pyngrok